chord-part-2/scripts/requirements.txt
requests
msgpack-rpc-python
